Changed ID parameter type from int to Guid
The ID parameter type in various methods and classes has been changed from `int` to `Guid`. This change has been applied to the `Get`, `Put`, and `Delete` methods in the `ArrivalsController`, `DeparturesController`, `DriversController`, `ParksController`, `TagsController`, and `VehiclesController` classes, as well as to several command and query classes. This implies that the system now identifies Arrivals, Departures, Drivers, Parks, Tags, and Vehicles using a `Guid` instead of an `int`, for a more unique identifier or to match changes in the underlying database schema.